backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
pydantic[email]==2.10.*
python-dotenv==1.1.*
bcrypt==4.3.*
PyJWT==2.10.*
prometheus-client==0.21.*
httpx==0.28.*
pytest==8.3.*
pytest-asyncio==0.25.*
